Move is_namedtuple_instance to utils and make it even more strict
